Refactor args out of principle module
